[ALPS07623055] UART Hub: Initialize variable
initialize variable

MTK-Commit-Id: 533007d99c2544ea4427c1214fa2d83d1ef6d06d

CR-Id: ALPS07623055
Feature: UART Hub

+++ drivers/misc/mediatek/uarthub/common/uarthub_drv_core.h
@@ -27,10 +27,10 @@ struct debug_info_ctrl {
 };
 
 struct uarthub_gpio_base_addr {
-	unsigned long addr;
-	unsigned long mask;
-	unsigned long value;
-	unsigned long gpio_value;
+	unsigned int addr;
+	unsigned int mask;
+	unsigned int value;
+	unsigned int gpio_value;
 };
 
 struct uarthub_gpio_trx_info {
